% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dol.R
\name{dol_disclosure_all}
\alias{dol_disclosure_all}
\title{Download all sheets from a DOL disclosure XLSX file}
\usage{
dol_disclosure_all(url)
}
\arguments{
\item{url}{Full URL to the XLSX file}
}
\value{
tibble with all sheets stacked, with a sheet_name column
}
\description{
Download all sheets from a DOL disclosure XLSX file
}
